refactor(models): type common SponsorFlags as SponsorCoverage
The base-class property was a raw uint like the legacy common Flags
pattern; since SponsorCoverage already exists and there is no per-tx
shadowing point for a common field, expose the typed enum directly.
Wire format unchanged - System.Text.Json serializes the enum as its
numeric value (no global string-enum converter is registered).

+++ Tests/Xrpl.Tests/Integration/transactions/TestISponsorship.cs
@@ -115,7 +115,7 @@ public class TestISponsorship
             Destination = destination.ClassicAddress,
             Amount = new Currency { ValueAsXrp = 1m },
             Sponsor = sponsor.ClassicAddress,
-            SponsorFlags = (uint)SponsorCoverage.spfSponsorFee,
+            SponsorFlags = SponsorCoverage.spfSponsorFee,
         };
         payment = await client.Autofill(payment);
